# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UserRole {
  role: string;
}

export interface UserRegistration {
  address: string;
  name: string;
}

export interface User {
  address: string;
  name: string;
  role: string;
}

export interface UserRegistrationResponse {
  success: boolean;
  user: User;
}

export interface UserRoleAssignment {
  address: string;
  role: string;
}

export interface UserRoleAssignmentResponse {
  success: boolean;
  user: User;
}