# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: user
    description: User operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /user/{address}/role:
    get:
      operationId: getUserRole
      tags: [user]
      summary: Get a user's role
      description: Returns the role associated with the provided address
      parameters:
        - name: address
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserRole"
        "404":
          description: User not found
  /user/register:
    post:
      operationId: registerUser
      tags: [user]
      summary: Register a user
      description: Stores a user in memory for the current server process
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegistration"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserRegistrationResponse"
        "400":
          description: Invalid registration data
  /user/assign-role:
    post:
      operationId: assignUserRole
      tags: [user]
      summary: Assign a user's role
      description: Updates the role for a registered user in the current server process
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRoleAssignment"
      responses:
        "200":
          description: User role assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserRoleAssignmentResponse"
        "400":
          description: Invalid role assignment data
        "404":
          description: User not found
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    UserRole:
      type: object
      properties:
        role:
          type: string
      required:
        - role
    UserRegistration:
      type: object
      properties:
        address:
          type: string
        name:
          type: string
      required:
        - address
        - name
    User:
      type: object
      properties:
        address:
          type: string
        name:
          type: string
        role:
          type: string
      required:
        - address
        - name
        - role
    UserRegistrationResponse:
      type: object
      properties:
        success:
          type: boolean
        user:
          $ref: "#/components/schemas/User"
      required:
        - success
        - user
    UserRoleAssignment:
      type: object
      properties:
        address:
          type: string
        role:
          type: string
      required:
        - address
        - role
    UserRoleAssignmentResponse:
      type: object
      properties:
        success:
          type: boolean
        user:
          $ref: "#/components/schemas/User"
      required:
        - success
        - user