# lib/api-zod/src/generated/types/userRegistrationResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { User } from './user';

export interface UserRegistrationResponse {
  success: boolean;
  user: User;
}