# lib/api-zod/src/generated/types/userRoleAssignmentResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { User } from './user';

export interface UserRoleAssignmentResponse {
  success: boolean;
  user: User;
}